Queue showing table view
