+++ Q/Q/Base.lproj/Main.storyboard
@@ -538,6 +538,32 @@
                                 <rect key="frame" x="20" y="112" width="560" height="480"/>
                                 <animations/>
                                 <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                                <prototypes>
+                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="blue" hidesAccessoryWhenEditing="NO" indentationLevel="1" indentationWidth="0.0" reuseIdentifier="cell" textLabel="K32-t2-496" style="IBUITableViewCellStyleDefault" id="rhG-sJ-QMp">
+                                        <rect key="frame" x="0.0" y="28" width="560" height="44"/>
+                                        <autoresizingMask key="autoresizingMask"/>
+                                        <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="rhG-sJ-QMp" id="4Of-UH-4VO">
+                                            <rect key="frame" x="0.0" y="0.0" width="560" height="43.5"/>
+                                            <autoresizingMask key="autoresizingMask"/>
+                                            <subviews>
+                                                <label opaque="NO" multipleTouchEnabled="YES" contentMode="left" text="Title" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" id="K32-t2-496">
+                                                    <rect key="frame" x="15" y="0.0" width="530" height="43.5"/>
+                                                    <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                                                    <animations/>
+                                                    <fontDescription key="fontDescription" type="system" pointSize="16"/>
+                                                    <color key="textColor" cocoaTouchSystemColor="darkTextColor"/>
+                                                    <nil key="highlightedColor"/>
+                                                </label>
+                                            </subviews>
+                                            <animations/>
+                                        </tableViewCellContentView>
+                                        <animations/>
+                                    </tableViewCell>
+                                </prototypes>
+                                <connections>
+                                    <outlet property="dataSource" destination="qPR-ck-kAc" id="THa-hj-m1K"/>
+                                    <outlet property="delegate" destination="qPR-ck-kAc" id="6Us-ac-yAm"/>
+                                </connections>
                             </tableView>
                         </subviews>
                         <animations/>
@@ -553,6 +579,9 @@
                             <constraint firstAttribute="trailingMargin" secondItem="QuM-hx-PEU" secondAttribute="trailing" id="w6R-qy-V7z"/>
                         </constraints>
                     </view>
+                    <connections>
+                        <outlet property="tableView" destination="QuM-hx-PEU" id="tgI-cv-zxz"/>
+                    </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="sF1-1g-09J" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
