+++ Q/Q/ProfessorQueueViewController.swift
@@ -15,7 +15,13 @@ class ProfessorQueueViewController: UIViewController {
 
     let user = PFUser.currentUser()!
     
-    var queueList = [Person]()
+    @IBOutlet weak var tableView: UITableView!
+    
+    var queueList = [Person]() {
+        didSet {
+            tableView.reloadData()
+        }
+    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -24,23 +30,56 @@ class ProfessorQueueViewController: UIViewController {
     }
     
     func loadQueueList() {
-        
-        guard let usernames = ParseUtility.userNamesOfQueueWithObjectId(queueId) else {
-            print("Error getting usernames")
-            return
-        }
+        let query = PFQuery(className: "Queue").whereKey("objectId", equalTo: queueId).includeKey("waitlist")
+        query.findObjectsInBackgroundWithBlock { queueWaitlist, error  in
             
-        //Iterate through all users and create the Model queueList for the TableView
+            guard error == nil else {
+                self.displayErrorString(error)
+                return
+            }
+            
+            guard let selectedQueue = queueWaitlist?.first else {
+                self.displayAlert("Error", message: "Could not get queue for some reason")
+                return
+            }
+            
+            guard let waitlist = selectedQueue["waitlist"] as? [String] else {
+                self.displayAlert("Error", message: "Could not get waitlist from selected queue")
+                return
+            }
+            
+            //Iterate through all users and create the Model queueList for the TableView
+            for username in waitlist {
                 
-    }
-        
-    }
+                let userQuery = PFUser.query()?.whereKey("username", equalTo: username)
+                userQuery?.findObjectsInBackgroundWithBlock { user, error in
+                    
+                    guard error == nil else {
+                        self.displayErrorString(error)
+                        return
+                    }
 
+                    
+                    guard let user = user?.first else {
+                        self.displayAlert("Error", message: "User not found")
+                        return
+                    }
+                    
+                    self.queueList.append(Person(lastName: user["lastName"] as! String,
+                        firstName: user["firstName"] as! String,
+                        userName: user["username"] as! String))
+                }
+            }
+        }
+    }
+}
 
 extension ProfessorQueueViewController: UITableViewDataSource, UITableViewDelegate {
     
     func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCellWithIdentifier("cell", forIndexPath: indexPath)
+        let person = queueList[indexPath.row]
+        cell.textLabel?.text = "\(person.firstName) \(person.lastName)"
         
         return cell
